requirements.txt
asgiref==3.7.2
Django==3.2.15
djangorestframework_simplejwt
django-cors-headers==4.0.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.4
typing-extensions==4.6.3
